認証関係を作り直した

+++ api/app/controllers/api/v1/users_controller.rb
@@ -1,27 +1,37 @@
 class Api::V1::UsersController < ApplicationController
-  include Firebase::FirebaseAuth
+  skip_before_action :authenticate_with_firebase_id_token!, only: %i[create]
 
   def show
-    request_certificates
-    raise ArgumentError, 'BadRequest Parameter' if payload.blank?
-
-    user = User.find_by(uid: params[:uid])
-    render json: user.as_json(include: :games)
+    user = User.find(params[:id])
+    render json: user.as_json(include: :games), status: :ok
   end
 
   def create
-    request_certificates
+    fetch_certificates
     raise ArgumentError, 'BadRequest Parameter' if payload.blank?
 
-    user = User.find_or_create_by(uid: uid_from_payload) do |u|
-      u.name = name_from_payload
-      u.avatar_url = avatar_url_from_payload
+    name = name_from_payload || 'anonymous'
+
+    user = User.find_or_initialize_by(uid: uid_from_payload)
+    user.assign_attributes(
+      name:,
+      avatar_url: avatar_url_from_payload
+    )
+
+    if user.save
+      render json: user.as_json(include: :games), status: :ok
+    else
+      render json: { status: 'unprocessable_entity' }, status: :unprocessable_entity
     end
-    render json: user.as_json(include: :games)
   end
 
   def destroy
     user = User.find(params[:id])
-    user.destroy
+    if user == current_user
+      user.destroy
+      render json: { status: 'no_content' }, status: :no_content
+    else
+      render json: { status: 'forbidden' }, status: :forbidden
+    end
   end
 end
